Remove duplicates in MDP files
A user can provide a MDP file with the `-x` option. Yet, that MDP file
cannot overwrite options that are defined in the default MDP template.
Indeed, the option would the appear twice in the resulting MDP file,
which would cause Gromacs to crash.

This commit introduces option deduplication for MDP files. When an
option is defined several time, only the last definition is kept. Note
that the deduplication does not conserve the order, the comments, not
the empty lines.